Calculate quantified impact at a domain level

Given a dataframe of simulation results summarized at the domain level, create a summarization of the annual loss expected (ALE) with descriptors at the minimum, mean, maximum, standard deviation, and 95 levels.

calculate_domain_impact(domain_summary, domains)

Arguments

domain_summary
Domain-level summary of simulation results as returned by summarize_domains.

domains
Dataframe of all domains in scope.

Value

Dataframe.

Examples

data(domain_summary)
data(domains)
calculate_domain_impact(domain_summary, domains)

#> # A tibble: 14 x 7
#>    domain_id      min     mean      max       sd      var domain               
#>    <chr>        <dbl>    <dbl>    <dbl>    <dbl>    <dbl> <chr>                
#>  1 IM        1215142. 4324064. 9436348. 1278577. 6588796. Information Security~
#>  2 COMP       715491. 2554688. 5490534.  853930. 4091600. Compliance (COMP)    
#>  3 ISMP       597580. 2420754. 5435747.  800720. 3811909. Information Security~
#>  4 ORG         60771. 1860772. 4511901.  769359. 3284849. Organization of Info~
#>  5 AC          25967.  552781. 1606611.  269118. 1053280. Access Control (AC)  
#>  6 HR          24115.  573179. 1681663.  265608. 1010868. Human Resources Secu~
#>  7 OPS         46573.  226870.  963740.  140500.  512434. Communications and O~
#>  8 PRI             0.  133273.  904346.  140987.  393278. Privacy Practices (P~
#>  9 ASSET         426.   19369.   53798.    9009.   35491. Asset Management (AS~
#> 10 ADM             0.       0.       0.       0.       0. Information Systems ~
#> 11 BC              0.       0.       0.       0.       0. Business Continuity ~
#> 12 PHY             0.       0.       0.       0.       0. Physical and Environ~
#> 13 POL             0.       0.       0.       0.       0. Security Policy (POL)
#> 14 RISK            0.       0.       0.       0.       0. Risk Management (RIS~