Calculate control weaknesses on a domain level

For each domain, calculate the total number of threat events, loss events, percentage of vulnerability, mean threat capability exceedance, and mean difficult exceedance.

calculate_weak_domains(simulation_results, domains)

Arguments

simulation_results
Results of running the risk simulations.

domains
Domain titles and IDs as a dataframe.

Value

Dataframe.

Examples

data(simulation_results)
data(domains)
calculate_weak_domains(simulation_results, domains)

#> # A tibble: 14 x 7
#>    domain_id loss_events threat_events vuln  tc_exceedance diff_e~ domain      
#>    <chr>           <int>         <dbl> <chr> <chr>         <chr>   <chr>       
#>  1 ASSET            6232        6253   99.7% 36.7%         4.1%    Asset Manag~
#>  2 AC               6125        6189   99.0% 34.5%         25.1%   Access Cont~
#>  3 ISMP            25127       26218   95.8% 11.8%         2.2%    Information~
#>  4 ADM                43          64.0 67.2% 11.4%         6.0%    Information~
#>  5 IM              63306      109535   57.8% 15.3%         13.2%   Information~
#>  6 HR              12293       24500   50.2% 30.7%         15.2%   Human Resou~
#>  7 OPS             33842       71003   47.7% 46.8%         22.3%   Communicati~
#>  8 COMP            26461       58625   45.1% 16.7%         16.2%   Compliance ~
#>  9 ORG             23611       70811   33.3% 3.8%          6.4%    Organizatio~
#> 10 PRI              1571       32407   4.8%  3.8%          32.9%   Privacy Pra~
#> 11 PHY               157        6317   2.5%  20.5%         32.6%   Physical an~
#> 12 BC                 42        6189   0.7%  11.9%         50.8%   Business Co~
#> 13 RISK               52       12378   0.4%  13.5%         46.7%   Risk Manage~
#> 14 POL                42       32407   0.1%  6.5%          29.2%   Security Po~