Create initial template files

Given a base directory, copy the provided sample files into an inputs subdirectory. This makes the starter files available for customizing and data collection. The inputs directory will be created if not already present. Pre-existing files, if present, will not be overwritten. Also creates an empty results subdirectory as a default location for evaluator output.

create_templates(base_directory = "~/evaluator")

Arguments

base_directory
Parent directory under which to create starter files.

Value

A dataframe of the starter filenames, along with a flag on whether a file was copied.

Examples

# NOT RUN {
create_templates("~/evaluator")
# }