Encode qualitative data to quantitative parameters

Given an input of:

qualitative risk scenarios

qualitative capabilities

translation table from qualitative labels to quantitative parameters

encode_scenarios(scenarios, capabilities, mappings)

Arguments

scenarios
Qualitative risk scenarios dataframe.

capabilities
Qualitative program capabilities dataframe.

mappings
Qualitative to quantitative mapping dataframe.

Value

A dataframe of capabilities for the scenario and parameters for quantified simulation.

Details

Create a unified dataframe of quantitative scenarios ready for simulation.

Examples

data(qualitative_scenarios, capabilities, mappings)
encode_scenarios(qualitative_scenarios, capabilities, mappings)

#> # A tibble: 56 x 18
#>    scena… scenario tcomm doma… cont… diff… tef_l tef_ml tef_h tef_…  tc_l tc_ml
#>     <int> <chr>    <chr> <chr> <chr> <lis> <int>  <dbl> <int> <int> <int> <dbl>
#>  1      1 Inadequ… Orga… ORG   1, 5… <tib…    10 24.0      52     4    33  50.0
#>  2      2 Budget … Orga… ORG   1, 5… <tib…    10 24.0      52     4    33  50.0
#>  3      3 Competi… Orga… ORG   1, 5… <tib…     1  6.00     12     4    33  50.0
#>  4      4 Increas… Orga… ORG   1, 3… <tib…     1  6.00     12     4    33  50.0
#>  5      5 The inf… Orga… ORG   1, 7… <tib…     1  6.00     12     4    33  50.0
#>  6      6 Inconsi… Orga… POL   9, 1… <tib…    10 24.0      52     4     0  16.0
#>  7      7 Misuse … Orga… POL   9, 1… <tib…     1  6.00     12     4     0  16.0
#>  8      8 Noncomp… Orga… POL   9, 1… <tib…     0  0.100     1     4    33  50.0
#>  9      9 New or … Exte… COMP  8, 13 <tib…     1  6.00     12     4     0  16.0
#> 10     10 Interna… Orga… COMP  8, 13 <tib…    10 24.0      52     4    33  50.0
#> # ... with 46 more rows, and 6 more variables: tc_h <int>, tc_conf <int>, lm_l
#> #   <int>, lm_ml <dbl>, lm_h <int>, lm_conf <int>