Generate sample analysis report

Given a set of input files and summarized simulation results, create a skeleton risk analysis report. This report attempts to summarize the results of the analysis at a top level, using 95 metric, while also providing more detailed analysis at both a per-domain and per-scenario level.

generate_report(input_directory = "~/evaluator/inputs",
  results_directory = "~/evaluator/results", output_file,
  styles = NULL, include_header = NULL,
  focus_scenario_ids = c("RS-51", "RS-12"), format = "html",
  intermediates_dir = tempdir(), quiet = TRUE, ...)

Arguments

input_directory
Location of input files.

results_directory
Location of simulation results.

output_file
Full path to output file.

styles
Optional full path to CSS file to override default styles.

include_header
Optional full path to HTML to include in the HEAD section (HTML formats only).

focus_scenario_ids
IDs of scenarios of special interest.

format
Format to generate (html, pdf, word).

intermediates_dir
Location for intermediate knit files.

quiet
TRUE to suppress printing of pandoc output.

...
Any other parameters to pass straight to rmarkdown::render.

Value

Default return values of the rmarkdown::render function.

Details

This report includes several sections where an analyst will need to modify and fill in details for their specific organization. Of particular note is the Recommendations section, which will always need to be updated.

Examples

# NOT RUN {
generate_report("~/inputs", "~/results", "~/risk_report.html")
# }