Import capabilities from survey spreadsheet

Import capabilities from survey spreadsheet

import_capabilities(survey_file = system.file("survey", "survey.xlsx", package
  = "evaluator"), domains)

Arguments

survey_file
Path to survey XLSX file. If not supplied, a default sample file is used.

domains
Dataframe of domains and domain IDs.

Value

Extracted capabilities as a dataframe.

Examples

data(domains)
import_capabilities(domains = domains)

#> # A tibble: 60 x 4
#>       id domain_id capability                                           diff   
#>    <int> <chr>     <chr>                                                <chr>  
#>  1     1 ORG       A senior-level committee provides dedicated oversig~ 5 - Op~
#>  2     2 ISMP      An information security framework is established to~ 2 - Re~
#>  3     3 ORG       Delegated authority is established for incident res~ 2 - Re~
#>  4     4 ISMP      Information security activities are supported by tr~ 2 - Re~
#>  5     5 ORG       Information security activities between departments~ 4 - Ma~
#>  6     6 ISMP      Roles and responsibilities are established and clea~ 3 - De~
#>  7     7 ORG       Senior management's strategy and direction for the ~ 1 - In~
#>  8     8 COMP      The information security program supports the priva~ 2 - Re~
#>  9     9 POL       A designated group is set up to act as ombudsman fo~ 2 - Re~
#> 10    10 POL       External authorities and industry trends are monito~ 5 - Op~
#> # ... with 50 more rows