Scenario-level risk summary

A sample set of quantified information security risk exposure, summarized at the scenario level, for the demonstration (and artificial) MetroCare information security program.

mc_scenario_summary

Format

scenario_id

ID of the scenario

domain_id

domain id

control_description

control description

results

nested data frame of simulation results for the scenario

loss_events_mean

mean number of loss events

loss_events_median

median number of loss events

loss_events_min

minimum number of loss events

loss_events_max

maximum number of loss events

ale_median

median annual loss expected

ale_max

maximum annual loss expected

ale_var

value at risk, ale

sle_min

minimum single loss expectancy

sle_max

maximum single loss expectancy

sle_mean

mean single loss expectancy

sle_median

median single loss expectancy

mean_tc_exceedance

mean threat capability exceedance

mean_diff_exceedance

mean difficulty exceedance

mean_vuln

mean vulnerability of the scenario

Source

This is hypothetical information. Any similarity to any other entity is completely coincidental.