Run an OpenFAIR simulation at the TEF/TC/DIFF/LM levels with control strength taken as a mean across multiple controls.

Run an OpenFAIR model with parameters provided for TEF, TC, DIFF, and LM sampling. If there are multiple controls provided for a scenarios, the arithmetic mean (average) is taken across samples for all controls to get the effective control strength for a given simulation.

openfair_tef_tc_diff_lm(scenario, diff_estimates, n = 10^4,
  title = "Untitled", verbose = FALSE)

Arguments

scenario
List of tef_, tc_, and LM_ l/ml/h/conf parameters.

diff_estimates
Parameters for estimating the scenario difficulty.

n
Number of simulations to run. Defaults to 10,000.

title
Optional name of scenario. Defaults to 10,000.

verbose
Whether to print progress indicators.

Value

Dataframe of scenario name, threat_event count, loss_event count, mean TC and DIFF exceedance, and ALE samples.