Run simulations for all scenarios

Given a dataframe of quantitative scenarios, run an OpenFAIR Monte Carlo simulation for each scenario, returning a combined dataframe of all results.

run_simulations(scenario, simulation_count = 10000L,
  model = "openfair_tef_tc_diff_lm", verbose = FALSE)

Arguments

scenario
Quantitative scenarios.

simulation_count
Number of simulations for each scenario.

model
OpenFAIR model to use.

verbose
Whether verbose console output is requested.

Value

Dataframe of raw results.

Examples

data(quantitative_scenarios)
# run a single scenario in a trivial number (10) of trials
run_simulations(quantitative_scenarios[1, ], 10)

#> # A tibble: 10 x 13
#>    domain_id scenario_id simulation threat_events loss_events  vuln
#>    <chr>           <int>      <int>         <dbl>       <int> <dbl>
#>  1 ORG                 1          1          30.0           5 0.167
#>  2 ORG                 1          2          30.0          11 0.367
#>  3 ORG                 1          3          23.0           7 0.304
#>  4 ORG                 1          4          19.0          11 0.579
#>  5 ORG                 1          5          24.0           4 0.167
#>  6 ORG                 1          6          36.0          14 0.389
#>  7 ORG                 1          7          32.0           7 0.219
#>  8 ORG                 1          8          31.0           8 0.258
#>  9 ORG                 1          9          41.0          11 0.268
#> 10 ORG                 1         10          20.0           4 0.200
#> # ... with 7 more variables: mean_tc_exceedance <dbl>,
#> #   mean_diff_exceedance <dbl>, ale <dbl>, sle_max <dbl>, sle_min <dbl>,
#> #   sle_mean <dbl>, sle_median <dbl>