Calculate the difficulty presented by controls, given a function and parameters for that function

Calculate the difficulty presented by controls, given a function and parameters for that function

sample_diff(func = NULL, params = NULL)

Arguments

func
Function to use to simulate DIFF, defaults to rpert.

params
Optional parameters to pass to func.

Value

List containing type ("diff"), samples (as a vector), and details (as a list).

See also

Other OpenFAIR helpers: compare_tef_vuln, get_mean_control_strength, openfair_tef_tc_diff_lm, sample_lef, sample_lm, sample_tc, sample_tef, sample_vuln, select_loss_opportunities