Calculate the number of simulated threat event frequencies (TEF)

Calculate the number of simulated threat event frequencies (TEF)

sample_tef(func = NULL, params = NULL)

Arguments

func
Function to use to simulate TEF, defaults to rpert.

params
Optional parameters to pass to func.

Value

List containing type ("tef"), samples (as a vector), and details (as a list).

See also

Other OpenFAIR helpers: compare_tef_vuln, get_mean_control_strength, openfair_tef_tc_diff_lm, sample_diff, sample_lef, sample_lm, sample_tc, sample_vuln, select_loss_opportunities