Create domain-level summary of simulation results

Given a dataframe of raw results from run_simulations, summarize the individual results at a per-domain level. This domain-level summary is a useful data structure for aggregate reporting.

summarize_domains(simulation_results, domains)

Arguments

simulation_results
Simulation results dataframe.

domains
Domain mappings dataframe.

Value

Simulation results summarized by domain

Details

Summary stats created include:

Mean/Min/Max/Median are calculated for loss events

Median/Max/VaR are calculated for annual loss expected (ALE)

Mean/Median/Max/Min are calculated for single loss expected (SLE)

Mean percentage of threat capability exceeding difficulty on successful threat events

Mean percentage of difficulty exceeding threat capability on defended events

Vulnerability percentage

Z-score of ALE (outliers flagged as 2 >= z-score)

Examples

data(simulation_results)
data(domains)
summarize_domains(simulation_results, domains)

#> # A tibble: 14,000 x 4
#> # Groups:   domain_id [14]
#>    domain_id domain         simulation      ale
#>    <chr>     <chr>               <int>    <dbl>
#>  1 AC        Access Control          1  657398.
#>  2 AC        Access Control          2  961149.
#>  3 AC        Access Control          3 1201174.
#>  4 AC        Access Control          4  522665.
#>  5 AC        Access Control          5  209271.
#>  6 AC        Access Control          6  510432.
#>  7 AC        Access Control          7  809112.
#>  8 AC        Access Control          8 1013984.
#>  9 AC        Access Control          9 1046355.
#> 10 AC        Access Control         10  827309.
#> # ... with 13,990 more rows