Create a summary of the simulation results for a single scenario

Given a dataframe of raw results from run_simulations, create summary statistics for the scenario. This is generally the most granular level of useful data for reporting and analysis (full simulation results are rarely directly helpful).

summarize_scenario(simulation_result)

summarize_scenarios(simulation_results)

Arguments

simulation_result
Results object for a single scenario.

simulation_results
Simulation results dataframe.

Value

Dataframe of summary statistics.

Details

Summary stats created include: * Mean/Min/Max/Median are calculated for loss events * Median/Max/VaR are calculated for annual loss expected (ALE) * Mean/Median/Max/Min are calculated for single loss expected (SLE) * Mean percentage of threat capability exceeding difficulty on successful threat events * Mean percentage of difficulty exceeding threat capability on defended events * Vulnerability percentage

Examples

data(mc_simulation_results)
# summarize a single scenario
summarize_scenario(mc_simulation_results[[1, "results"]])

#> # A tibble: 1 x 14
#>   loss_events_mean loss_events_med… loss_events_min loss_events_max ale_median
#>              <dbl>            <dbl>           <dbl>           <dbl>      <dbl>
#> 1             7.48                7               0              19    682568.
#> # … with 9 more variables: ale_max <dbl>, ale_var <dbl>, sle_mean <dbl>,
#> #   sle_median <dbl>, sle_min <dbl>, sle_max <dbl>, mean_tc_exceedance <dbl>,
#> #   mean_diff_exceedance <dbl>, mean_vuln <dbl>

# summarize all scenarios in a data frame
data(mc_simulation_results)
summarize_scenarios(mc_simulation_results)

#> # A tibble: 56 x 18
#>    scenario_id domain_id control_descrip… results loss_events_mean
#>    <chr>       <chr>     <list>           <list>             <dbl>
#>  1 RS-01       ORG       <list [7]>       <tibbl…            7.48 
#>  2 RS-02       ORG       <list [7]>       <tibbl…            7.48 
#>  3 RS-03       ORG       <list [7]>       <tibbl…            1.72 
#>  4 RS-04       ORG       <list [8]>       <tibbl…            2.73 
#>  5 RS-05       ORG       <list [5]>       <tibbl…            4.21 
#>  6 RS-06       POL       <list [4]>       <tibbl…            0    
#>  7 RS-07       POL       <list [4]>       <tibbl…            0    
#>  8 RS-08       POL       <list [4]>       <tibbl…            0.042
#>  9 RS-09       COMP      <list [2]>       <tibbl…            0    
#> 10 RS-10       COMP      <list [2]>       <tibbl…            1.90 
#> # … with 46 more rows, and 13 more variables: loss_events_median <dbl>,
#> #   loss_events_min <dbl>, loss_events_max <dbl>, ale_median <dbl>,
#> #   ale_max <dbl>, ale_var <dbl>, sle_mean <dbl>, sle_median <dbl>,
#> #   sle_min <dbl>, sle_max <dbl>, mean_tc_exceedance <dbl>,
#> #   mean_diff_exceedance <dbl>, mean_vuln <dbl>