Calculate quantified impact at a domain level

Given a dataframe of simulation results summarized at the domain level, create a summarization of the annual loss expected (ALE) with descriptors at the minimum, mean, maximum, standard deviation, and 95 levels.

calculate_domain_impact(domain_summary, domains)

Arguments

domain_summary
Domain-level summary of simulation results as returned by summarize_domains.

domains
Dataframe of all domains in scope.

Value

Dataframe.

Examples

data(domain_summary)
data(domains)
calculate_domain_impact(domain_summary, domains)

#> # A tibble: 14 x 7
#>    domain_id      min     mean      max       sd      var domain               
#>    <chr>        <dbl>    <dbl>    <dbl>    <dbl>    <dbl> <chr>                
#>  1 IM        1513809. 4465097. 9367780. 1302761. 6691612. Information Security…
#>  2 COMP       655382. 2607239. 5676333.  859137. 4176184. Compliance (COMP)    
#>  3 ISMP       663743. 2474047. 5468451.  807195. 3963943. Information Security…
#>  4 ORG        231313. 1912130. 5295248.  782056. 3357335. Organization of Info…
#>  5 PHY             0   165788. 3898004.  654598. 2014140. Physical and Environ…
#>  6 HR          44730.  622980. 1855731.  275654. 1120515. Human Resources Secu…
#>  7 AC          15054.  599813. 1450198.  266695. 1083095. Access Control (AC)  
#>  8 OPS         43936.  251152. 1151143.  148927.  533819. Communications and O…
#>  9 PRI             0   159156.  786643.  147745.  436371. Privacy Practices (P…
#> 10 ASSET        1138.   27463.  321407.   32002.   58655. Asset Management (AS…
#> 11 RISK            0     5762.  321986.   30204.   15102. Risk Management (RIS…
#> 12 ADM             0     4741.  304163.   27386.       0  Information Systems …
#> 13 BC              0   102050. 4091811.  508817.       0  Business Continuity …
#> 14 POL             0      162.    8246.     939.       0  Security Policy (POL)