Calculate quantified impact at a domain level

Given a dataframe of simulation results summarized at the domain level, create a summarization of the annual loss expected (ALE) with descriptors at the minimum, mean, maximum, standard deviation, and 95 levels.

calculate_domain_impact(domain_summary, domains)

Arguments

domain_summary
Domain-level summary of run_simulation results.

domains
Dataframe of all domains in scope.

Value

Dataframe.

Examples

data(domain_summary)
data(domains)
calculate_domain_impact(domain_summary, domains)

#> # A tibble: 14 x 7
#>    domain_id     min    mean     max      sd     var domain                    
#>    <chr>       <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <chr>                     
#>  1 IM        1513809 4465097 9367780 1302761 6691612 Information Security Inci~
#>  2 COMP       743762 2598394 6145570  869087 4077983 Compliance (COMP)         
#>  3 ISMP       663743 2474047 5468451  807195 3963943 Information Security Mana~
#>  4 ORG        280905 1907368 5326646  758155 3237394 Organization of Informati~
#>  5 PHY             0  165788 3898004  654598 2014140 Physical and Environmenta~
#>  6 HR          44730  622980 1855731  275654 1120515 Human Resources Security ~
#>  7 AC          15054  599813 1450198  266695 1083095 Access Control (AC)       
#>  8 OPS         43936  251152 1151143  148927  533819 Communications and Operat~
#>  9 PRI             0  159156  786643  147745  436371 Privacy Practices (PRI)   
#> 10 ASSET        1138   27463  321407   32002   58655 Asset Management (ASSET)  
#> 11 RISK            0    5762  321986   30204   15102 Risk Management (RISK)    
#> 12 ADM             0    4741  304163   27386       0 Information Systems Acqui~
#> 13 BC              0  102050 4091811  508817       0 Business Continuity Manag~
#> 14 POL             0     169    8246     937       0 Security Policy (POL)