Encode qualitative data to quantitative parameters

Given an input of:

qualitative risk scenarios

qualitative capabilities

translation table from qualitative labels to quantitative parameters

encode_scenarios(scenarios, capabilities, mappings)

Arguments

scenarios
Qualitative risk scenarios dataframe.

capabilities
Qualitative program capabilities dataframe.

mappings
Qualitative to quantitative mapping dataframe.

Value

A dataframe of capabilities for the scenario and parameters for quantified simulation.

Details

Create a unified dataframe of quantitative scenarios ready for simulation.

Examples

data(mc_qualitative_scenarios, mc_capabilities, mc_mappings)
encode_scenarios(mc_qualitative_scenarios, mc_capabilities, mc_mappings)

#> # A tibble: 56 x 6
#>    scenario_id scenario_description  tcomm  domain_id control_descrip… scenario 
#>    <chr>       <chr>                 <chr>  <chr>     <list>           <list>   
#>  1 1           Inadequate human res… Techn… ORG       <list [7]>       <S3: tid…
#>  2 10          Internal workforce m… Organ… COMP      <list [2]>       <S3: tid…
#>  3 11          External attackers l… Exter… COMP      <list [2]>       <S3: tid…
#>  4 12          External auditors fi… Exter… COMP      <list [2]>       <S3: tid…
#>  5 13          Inability to accurat… Techn… RISK      <list [4]>       <S3: tid…
#>  6 14          Lack of risk managem… Techn… RISK      <list [4]>       <S3: tid…
#>  7 15          Inability to underst… Exter… RISK      <list [4]>       <S3: tid…
#>  8 16          Security failure or … Organ… RISK      <list [4]>       <S3: tid…
#>  9 17          Internal user inappr… Organ… PRI       <list [4]>       <S3: tid…
#> 10 18          Systemic noncomplian… Regul… PRI       <list [4]>       <S3: tid…
#> # … with 46 more rows