Run an OpenFAIR simulation at the TEF/TC/DIFF/LM levels

Run an OpenFAIR model with parameters provided for TEF, TC, DIFF, and LM sampling. If there are multiple controls provided for a scenarios, the arithmetic mean (average) is taken across samples for all controls to get the effective control strength for a given simulation.

openfair_tef_tc_diff_lm(scenario, n = 10^4, verbose = FALSE)

Arguments

scenario
A evaluator_scen object

n
Number of simulations to run.

verbose
Whether to print progress indicators.

Value

Dataframe of scenario name, threat_event count, loss_event count, mean TC and DIFF exceedance, and ALE samples.

See also

Examples

data(quantitative_scenarios)
scenario <- quantitative_scenarios[[1, "scenario"]]
openfair_tef_tc_diff_lm(scenario, 10)

#> # A tibble: 10 x 11
#>    simulation threat_events loss_events  vuln mean_tc_exceeda… mean_diff_excee…
#>         <int>         <int>       <int> <dbl>            <dbl>            <dbl>
#>  1          1            30           9 0.3               3.78             7.86
#>  2          2            30          11 0.367             3.48             6.51
#>  3          3            23           6 0.261             4.16             8.22
#>  4          4            19           8 0.421             2.48             6.48
#>  5          5            24           4 0.167             3.07             7.06
#>  6          6            36           7 0.194             4.20             4.98
#>  7          7            32           6 0.188             1.90             4.93
#>  8          8            31           8 0.258             1.52             6.83
#>  9          9            41           8 0.195             3.86             7.13
#> 10         10            20           4 0.2               4.00             8.72
#> # … with 5 more variables: ale <dbl>, sle_max <dbl>, sle_min <dbl>,
#> #   sle_mean <dbl>, sle_median <dbl>