Run an OpenFAIR simulation at the TEF/TC/DIFF/LM levels

Run an OpenFAIR model with parameters provided for TEF, TC, DIFF, and LM sampling. If there are multiple controls provided for a scenarios, the arithmetic mean (average) is taken across samples for all controls to get the effective control strength for a given simulation.

openfair_tef_tc_diff_lm(scenario, n = 10^4, title = "Untitled",
  verbose = FALSE)

Arguments

scenario
List of TEF, TC, and LM parameters.

n
Number of simulations to run.

title
Optional name of scenario.

verbose
Whether to print progress indicators.

Value

Dataframe of scenario name, threat_event count, loss_event count, mean TC and DIFF exceedance, and ALE samples.

See also

Examples

data(quantitative_scenarios)
scenario <- quantitative_scenarios[1, ]
openfair_tef_tc_diff_lm(scenario, 10)

#> # A tibble: 10 x 12
#>    title simulation threat_events loss_events  vuln mean_tc_exceeda…
#>    <chr>      <int>         <int>       <int> <dbl>            <dbl>
#>  1 Unti…          1            30           9 0.3               3.78
#>  2 Unti…          2            30          11 0.367             3.48
#>  3 Unti…          3            23           6 0.261             4.16
#>  4 Unti…          4            19           8 0.421             2.48
#>  5 Unti…          5            24           4 0.167             3.07
#>  6 Unti…          6            36           7 0.194             4.20
#>  7 Unti…          7            32           6 0.188             1.90
#>  8 Unti…          8            31           8 0.258             1.52
#>  9 Unti…          9            41           8 0.195             3.86
#> 10 Unti…         10            20           4 0.2               4.00
#> # ... with 6 more variables: mean_diff_exceedance <dbl>, ale <dbl>,
#> #   sle_max <dbl>, sle_min <dbl>, sle_mean <dbl>, sle_median <dbl>