Run an OpenFAIR simulation at the TEF/TC/DIFF/LM levels

Run an OpenFAIR model with parameters provided for TEF, TC, DIFF, and LM sampling. If there are multiple controls provided for a scenarios, the arithmetic mean (average) is taken across samples for all controls to get the effective control strength for a given simulation.

openfair_tef_tc_diff_lm(scenario, diff_estimates, n = 10^4,
  title = "Untitled", verbose = FALSE)

Arguments

scenario
List of tef_, tc_, and LM_ l/ml/h/conf parameters.

diff_estimates
Parameters for estimating the scenario difficulty.

n
Number of simulations to run.

title
Optional name of scenario.

verbose
Whether to print progress indicators.

Value

Dataframe of scenario name, threat_event count, loss_event count, mean TC and DIFF exceedance, and ALE samples.

See also

Other OpenFAIR helpers: compare_tef_vuln, get_mean_control_strength, sample_diff, sample_lef, sample_lm, sample_tc, sample_tef, sample_vuln, select_loss_opportunities

Examples

data(quantitative_scenarios)
scenario <- quantitative_scenarios[1, ]
controls <- scenario[[1, "diff_params"]]
openfair_tef_tc_diff_lm(scenario, controls, 10)

#> # A tibble: 10 x 12
#>    title    simulation threat_events loss_events  vuln mean_tc_exceedance
#>    <chr>         <int>         <dbl>       <int> <dbl>              <dbl>
#>  1 Untitled          1           30.           5 0.167               4.63
#>  2 Untitled          2           30.          11 0.367               3.23
#>  3 Untitled          3           23.           7 0.304               2.49
#>  4 Untitled          4           19.          11 0.579               3.30
#>  5 Untitled          5           24.           4 0.167               2.81
#>  6 Untitled          6           36.          14 0.389               2.55
#>  7 Untitled          7           32.           7 0.219               4.05
#>  8 Untitled          8           31.           8 0.258               2.10
#>  9 Untitled          9           41.          11 0.268               2.99
#> 10 Untitled         10           20.           4 0.200               4.59
#> # ... with 6 more variables: mean_diff_exceedance <dbl>, ale <dbl>,
#> #   sle_max <dbl>, sle_min <dbl>, sle_mean <dbl>, sle_median <dbl>