Run simulations for all scenarios

Given a dataframe of quantitative scenarios, run an OpenFAIR Monte Carlo simulation for each scenario, returning a combined dataframe of all results.

run_simulations(scenario, simulation_count = 10000L,
  model = "openfair_tef_tc_diff_lm")

Arguments

scenario
Quantitative scenarios.

simulation_count
Number of simulations for each scenario.

model
OpenFAIR model to use.

Value

Dataframe of raw results.

Examples

data(quantitative_scenarios)
# run a single scenario in a trivial number (10) of trials
run_simulations(quantitative_scenarios[1, ], 10)

#> # A tibble: 10 x 13
#>    domai… scena… simu… thre… loss…  vuln mean… mean…    ale sle_m… sle_… sle_m…
#>    <chr>   <int> <int> <dbl> <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl>  <dbl>
#>  1 ORG         1     1  30.0     5 0.167  4.63  6.90 3.54e⁵ 155138 30123  70701
#>  2 ORG         1     2  30.0    11 0.367  3.23  7.23 1.66e⁶ 327020 27705 150734
#>  3 ORG         1     3  23.0     7 0.304  2.49  8.42 6.67e⁵ 163373 33813  95300
#>  4 ORG         1     4  19.0    11 0.579  3.30  6.04 1.03e⁶ 287868 10307  93932
#>  5 ORG         1     5  24.0     4 0.167  2.81  7.59 2.78e⁵ 119325 42075  69579
#>  6 ORG         1     6  36.0    14 0.389  2.55  6.33 1.61e⁶ 199345 18198 115312
#>  7 ORG         1     7  32.0     7 0.219  4.05  5.80 5.55e⁵ 173387 39962  79336
#>  8 ORG         1     8  31.0     8 0.258  2.10  7.11 5.72e⁵ 196275 20422  71560
#>  9 ORG         1     9  41.0    11 0.268  2.99  6.60 1.10e⁶ 258991 32951  99884
#> 10 ORG         1    10  20.0     4 0.200  4.59  8.55 4.03e⁵ 151144 23716 100775
#> # ... with 1 more variable: sle_median <dbl>