Create domain-level summary of simulation results

Given a dataframe of raw results from run_simulations, summarize the individual results at a per-domain level. This domain-level summary is a useful data structure for aggregate reporting.

summarize_domains(simulation_results)

Arguments

simulation_results
Simulation results dataframe.

Value

Simulation results summarized by domain

Details

Summary stats created include:

Mean/Min/Max/Median are calculated for loss events

Median/Max/VaR are calculated for annual loss expected (ALE)

Mean/Median/Max/Min are calculated for single loss expected (SLE)

Mean percentage of threat capability exceeding difficulty on successful threat events

Mean percentage of difficulty exceeding threat capability on defended events

Vulnerability percentage

Z-score of ALE (outliers flagged as 2 >= z-score)

Examples

data(simulation_results)
summarize_domains(simulation_results)

#> # A tibble: 14 x 13
#>    domain_id  ale_min ale_median ale_mean  ale_max   ale_sd  ale_var
#>    <chr>        <dbl>      <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1 AC          15054.    579625.  599813. 1450198.  266695. 1083095.
#>  2 ADM             0          0     4741.  304163.   27386.       0 
#>  3 ASSET        1138.     21127.   27463.  321407.   32002.   58655.
#>  4 BC              0          0   102050. 4091811.  508817.       0 
#>  5 COMP       655382.   2517171. 2607239. 5676333.  859137. 4176184.
#>  6 HR          44730.    584575.  622980. 1855731.  275654. 1120515.
#>  7 IM        1513809.   4322443. 4465097. 9367780. 1302761. 6691612.
#>  8 ISMP       663743.   2410077. 2474047. 5468451.  807195. 3963943.
#>  9 OPS         43936.    219246.  251152. 1151143.  148927.  533819.
#> 10 ORG        231313.   1819913. 1912130. 5295248.  782056. 3357335.
#> 11 PHY             0          0   165788. 3898004.  654598. 2014140.
#> 12 POL             0          0      162.    8246.     939.       0 
#> 13 PRI             0     131708.  159156.  786643.  147745.  436371.
#> 14 RISK            0          0     5762.  321986.   30204.   15102.
#> # ... with 6 more variables: mean_threat_events <dbl>, mean_loss_events <dbl>,
#> #   mean_avoided_events <dbl>, mean_tc_exceedance <dbl>,
#> #   mean_diff_exceedance <dbl>, mean_vuln <dbl>